refactor(ffi): widen scf_atom.Z / CAtom.atomic_number from int to double
Fractional external point charges (e.g. QM/MM partial charges) need a
non-integer Z; libint2's set_params already takes double internally, only
the C ABI struct was narrowing it. No behavior change for existing
integer-charge callers.

Also fixes the one other .Z read site in shim.cc (scf_basis_create's
li_atoms[a].atomic_number assignment, which maps into libint2::Atom, an
external struct with an int atomic_number field used only for basis-set/
shell element lookup) with an explicit static_cast<int> now that atoms[a].Z
is a double — this preserves prior behavior for all real (integer) nuclei
while making the truncation explicit rather than an implicit narrowing
conversion.

+++ crates/ferric-integrals/src/basis_bridge.rs
@@ -59,7 +59,7 @@ impl PreparedBasis {
                 }
             };
             CAtom {
-                atomic_number: z_eff as c_int, x: a.x, y: a.y, z: a.zpos,
+                atomic_number: z_eff as f64, x: a.x, y: a.y, z: a.zpos,
             }
         }).collect();
